+++ CHANGELOG.md
@@ -4,5 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.1] - 2017-03-17
+Refactored to use forEach for promise.all pattern.
+Turns out recursion and promises are friends of race condtions.
+
 ## [1.0.0] - 2017-03-17
 First Version